refactor: 예약, 시간, 테마 패키지 이동
- entity -> domain으로 이동 (예약이 상태를 가진 객체가 되었기 때문)
