+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,7 +1,7 @@
 package roomescape.dto;
 
 import java.time.LocalDate;
-import roomescape.entity.Reservation;
+import roomescape.domain.Reservation;
 
 public record ReservationResponse(
         Long id,
