+++ src/main/java/roomescape/dto/ThemeResponse.java
@@ -1,6 +1,6 @@
 package roomescape.dto;
 
-import roomescape.entity.Theme;
+import roomescape.domain.Theme;
 
 public record ThemeResponse(
         Long id,
