+++ src/main/java/roomescape/infra/JdbcThemeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.entity.Theme;
+import roomescape.domain.Theme;
 import roomescape.repository.ThemeRepository;
 
 @Repository
