+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -3,7 +3,7 @@ package roomescape.repository;
 import java.util.List;
 import java.util.Optional;
 import roomescape.common.exception.ResourceNotFoundException;
-import roomescape.entity.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 public interface ReservationTimeRepository {
