+++ src/main/java/roomescape/service/ThemeService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.common.exception.InvalidDeleteException;
+import roomescape.domain.Theme;
 import roomescape.dto.ThemeRequest;
-import roomescape.entity.Theme;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
