+++ src/test/java/roomescape/infra/JdbcThemeRepositoryTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.entity.Theme;
+import roomescape.domain.Theme;
 import roomescape.repository.ThemeRepository;
 
 @JdbcTest
